backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
pillow
numpy
opencv-python-headless
pydicom
reportlab
jinja2
python-dateutil
boto3
python-dotenv
passlib[bcrypt]
torchxrayvision
scikit-image
python-jose[cryptography]
transformers
timm
einops
scikit-learn
scipy
neurokit2
pandas
